requirements.txt
asgiref==3.3.1
attrs==21.4.0
certifi==2021.5.30
charset-normalizer==2.0.4
cmd2==2.3.3
distlib==0.3.4
Django==3.1.4
django-ckeditor==6.0.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-phone-field==1.8.1
django-project-template-loader==0.1.2
django-redirect-to-non-www==0.1.1
django-robots==5.0
django-sitemaps==2.0.1
filelock==3.4.2
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.10.1
jsons==1.6.1
lxml==4.7.1
Mypackage==0.1
Pillow==8.1.0
pipenv==2022.1.8
platformdirs==2.4.1
pyperclip==1.8.2
pytz==2020.5
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==4.0.1
typish==1.9.3
urllib3==1.26.6
virtualenv==20.13.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
zipp==3.7.0
